added bell icons and drawer for alert
