+++ src/layouts/topNav/topNav.scss
@@ -102,13 +102,6 @@
 		img {
 			cursor: pointer;
 		}
-		div {
-			position: relative;
-			.ellipse {
-				position: absolute;
-				right: 0;
-			}
-		}
 	}
 
 	.user_info {
