+++ src/utils/ConvertToDateTime.js
@@ -1,8 +1,6 @@
 import dayjs from 'dayjs';
-import utc from 'dayjs/plugin/utc';
-dayjs.extend(utc);
 const ConvertToDateTime = (dateString, format = 'YYYY-MM-DD') => {
-    return dateString && dayjs.utc(dateString).format(format);
+    return dateString && dayjs(dateString).format(format);
 };
 
 export default ConvertToDateTime;
